refactor: remediate JD Round 1 dependency violations and hidden bugs (377 tests)
- Move pure HTTP utilities (auth, json, cors, text, types) to infra/http/ so both
  transport/ and integrations/ can import them without sibling cross-imports
- Move validateHookToken to infra/http/auth; integrations/codex re-exports alias
  so transport/ tests no longer import from integrations/
- Move shouldWriteProjectState to infra/paths/; core/state/store re-exports it;
  infra/banner/writer imports from infra/paths directly (no core→infra upward dep)
- Extract NotificationService port to core/types/notification-service so
  integrations/opencode imports from core/ instead of notifications/ sibling
- Refactor createNotificationService to accept NotificationServiceDeps object;
  implementation iterates channels[] for fan-out on all three notification methods
- Store codexHandle in server/index; shutdown() follows spec order:
  integrations → http → tunnel → notifications.flush → clearState
- Fix asset-sanity.test.ts: readdirSync now recursive so all dashboard JS
  sub-directories are scanned for hardcoded PILOT_VERSION strings
- Remove empty zombie dirs: src/server/{hooks,services,util}
- Fix system.ts and routes.ts docblock path drift (server/dashboard → dashboard)
- Push generateVapid through PushService port; settings handler calls deps.push.generateVapid()
