+++ src/dashboard/__tests__/integration.test.ts
@@ -1,6 +1,6 @@
 // integration.test.ts — End-to-end HTTP integration tests for critical flows.
 // Tests run against a real Bun.serve instance with a fake OpenCode SDK client.
-// Mirrors the pattern in src/server/http/server.test.ts.
+// Mirrors the pattern in src/transport/http/__tests__/server.test.ts.
 
 import { describe, it, expect, beforeAll, afterAll } from "bun:test"
 import type { PluginInput } from "@opencode-ai/plugin"
@@ -8,8 +8,8 @@ import type { Config } from "../../server/config"
 import type { AuditLog } from "../../core/audit/log"
 import { createEventBus } from "../../core/events/bus"
 import { createPermissionQueue } from "../../core/permissions/queue"
-import { createTelegramBot } from "../../notifications/channels/telegram/index"
-import { createPushService } from "../../notifications/channels/push/service"
+import { createTelegramChannel } from "../../notifications/pipeline"
+import { createPushService } from "../../notifications/pipeline"
 import type { Logger } from "../../infra/logger/index"
 import type { RouteDeps } from "../../transport/http/routes"
 import { createRemoteServer, type RemoteServer } from "../../transport/http/server"
@@ -120,7 +120,7 @@ function buildDeps(port: number): RouteDeps {
   const eventBus = createEventBus()
   const permissionQueue = createPermissionQueue(config.permissionTimeoutMs)
   const codexPermissionQueue = createPermissionQueue(config.codexPermissionTimeoutMs)
-  const telegram = createTelegramBot(null, permissionQueue, codexPermissionQueue)
+  const telegram = createTelegramChannel(null, permissionQueue, codexPermissionQueue)
   const audit = createNoopAudit()
   const logger = createNoopLogger()
   const push = createPushService({ config, audit, logger })
