+++ src/integrations/opencode/hooks/event.ts
@@ -1,5 +1,5 @@
 import type { PluginInput } from "@opencode-ai/plugin"
-import type { NotificationService } from "../../../notifications/pipeline"
+import type { NotificationService } from "../../../core/types/notification-service"
 
 /**
  * Handle all SDK events forwarded by OpenCode.
