+++ src/transport/http/middlewares/json.ts
@@ -1,32 +1,5 @@
-export interface ErrorBody {
-  error: {
-    code: string
-    message: string
-  }
-}
-
-/** Typed JSON response helper. */
-export function json<T>(
-  data: T,
-  status: number,
-  extraHeaders: Record<string, string> = {},
-): Response {
-  return new Response(JSON.stringify(data, null, 2), {
-    status,
-    headers: {
-      "Content-Type": "application/json",
-      ...extraHeaders,
-    },
-  })
-}
-
-/** Convenience for error responses. */
-export function jsonError(
-  code: string,
-  message: string,
-  status: number,
-  extraHeaders: Record<string, string> = {},
-): Response {
-  const body: ErrorBody = { error: { code, message } }
-  return json(body, status, extraHeaders)
-}
+// Re-exports from infra/http/json so consumers that import from this path
+// continue to work. The implementation lives in infra/ so integrations/ can
+// also import it without violating the sibling cross-import rule.
+export type { ErrorBody } from "../../../infra/http/json"
+export { json, jsonError } from "../../../infra/http/json"
