+++ src/transport/http/server.ts
@@ -61,40 +61,9 @@ export function checkBodySize(req: Request, maxBytes = MAX_REQUEST_BODY_BYTES): 
   return null
 }
 
-/**
- * Read a request body up to `maxBytes`, streaming chunk-by-chunk so that
- * chunked-encoded requests (no Content-Length) are also bounded.
- *
- * Returns the body text on success, or `null` if the body exceeds `maxBytes`.
- * Returns an empty string when the body is absent.
- *
- * Use this instead of bare `req.text()` in handlers that accept untrusted input
- * from external callers (e.g. Codex hook bridge) where Content-Length cannot
- * be trusted.
- */
-export async function readBoundedText(req: Request, maxBytes = MAX_REQUEST_BODY_BYTES): Promise<string | null> {
-  if (!req.body) return ""
-  const reader = req.body.getReader()
-  const decoder = new TextDecoder()
-  let total = 0
-  let result = ""
-  try {
-    while (true) {
-      const { done, value } = await reader.read()
-      if (done) break
-      total += value.byteLength
-      if (total > maxBytes) {
-        try { reader.cancel() } catch {}
-        return null
-      }
-      result += decoder.decode(value, { stream: true })
-    }
-    result += decoder.decode()
-  } finally {
-    try { reader.releaseLock() } catch {}
-  }
-  return result
-}
+// Re-export readBoundedText from infra so existing consumers that import from
+// this path continue to work. The implementation lives in infra/http/text.ts.
+export { readBoundedText } from "../../infra/http/text"
 
 export function createRemoteServer(deps: RouteDeps): RemoteServer {
   let server: ReturnType<typeof Bun.serve> | null = null
